initial working evaluation stack finds entry points into graph and then evaluates in depth first manner if possible backtracking where necessary.
still not yet uploading python lib and iron python .dlls att.

refactored position port method to port view, this is called on port start

refactored node creation into generic method command

problem around the design of generic base view<basemodel> it is difficult to store a matching base view field on the model since the generic type would be the type of model, can use a generic method but it's not ideal.