explicit execution is working at this point, unfortunately monodevelop again messed with line endings making it look like I changed many files.
core idea is to push delegates into python or any evaluation, these delegates just call the triggers on the nodes which in turn call further execution methods... I made an explicit activator evaluator that just grabs all nodes with no input execution triggers and calls them when the run button is pressed... to get observable execution need to call these delegates with generators.  I'd like to slow down execution so that one node executes per frame for now.

Other necessary next steps:
in the context of python, intermediate variables that we might use downstream are not available, for instance in example for loop :

for i in range(input1):
	calloutput(0)
	//use i to do something in another node downstream
calloutput(1)

i will not be available, to get this working will need to gather all outputs anytime any output trigger is called, this will update variable values just in time for any future use... in the simple case at least.

will need to consider if continuing to work on this implementation is worth the benefit -