made a fix to open dialog for opening any kind of file
modified eval method on custom node to use an evaluator, it now gathers execution data from single downstream input to custom node graph and immediately gathers exec data before evaluating it so that there is a current task in the queue.

added public method to inputexenode to force gather data to be regenerated

investigating similar problem on outputexecNode where caller is not set and output trigger on original wrapper cannot be found.